First Commit!
Got the blasted import to work finally! Can work on from home computer
now. Added comments to some of the files for future debugging. I don't
think we're running this code on the robot anyway - Noah reimaged the
robot with some stuff he found on the web, which I'm cool with. The fact
that it works is what stings... Will try to find what he used and upload
to github so we can work off of it. What's with the camera move class?
Are we going to be rotating it?